Refactor chat module: Remove old chat UI and configuration files, migrate chat handlers to new structure
- Deleted chat.js, config.js, and index.html files to remove the old chat UI implementation.
- Updated http.go to remove chat-related routes and handlers.
- Refactored websocket handling in http_ws.go to integrate new chat functionalities.
- Removed obsolete http_ws_llm.go file and its associated logic.
- Updated tests in http_ws_test.go to align with the new chat structure and functionalities.
