+++ mods/server/http_ws_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/gorilla/websocket"
 	"github.com/machbase/neo-server/v8/mods/eventbus"
+	"github.com/machbase/neo-server/v8/mods/server/chat"
 	"github.com/stretchr/testify/require"
 )
 
@@ -15,6 +16,21 @@ func TestWsLLMGetProviders(t *testing.T) {
 		params []interface{}
 		expect any
 	}{
+		{
+			method: "llmGetClaudeConfig",
+			params: nil,
+			expect: map[string]interface{}{
+				"key":        "your-key",
+				"max_tokens": float64(1024),
+			},
+		},
+		{
+			method: "llmGetOllamaConfig",
+			params: nil,
+			expect: map[string]interface{}{
+				"url": "http://127.0.0.1:11434",
+			},
+		},
 		{
 			method: "llmGetProviders",
 			params: nil,
@@ -42,7 +58,7 @@ func TestWsLLMGetProviders(t *testing.T) {
 	}
 
 	// Ensure we have some LLM providers loaded
-	useTestingLLMProviders()
+	chat.SetTesting(true)
 
 	at, _, err := jwtLogin("sys", "manager")
 	require.Nil(t, err)
@@ -54,34 +70,36 @@ func TestWsLLMGetProviders(t *testing.T) {
 	defer ws.Close()
 
 	for id, tc := range tests {
-		rpcReq := &eventbus.RPC{
-			Ver:    "2.0",
-			ID:     int64(id + 1),
-			Method: tc.method,
-			Params: tc.params,
-		}
-		req := eventbus.Event{Type: eventbus.EVT_RPC_REQ, Rpc: rpcReq}
-		err = ws.WriteJSON(req)
-		require.NoError(t, err)
+		t.Run(tc.method, func(t *testing.T) {
+			rpcReq := &eventbus.RPC{
+				Ver:    "2.0",
+				ID:     int64(id + 1),
+				Method: tc.method,
+				Params: tc.params,
+			}
+			req := eventbus.Event{Type: eventbus.EVT_RPC_REQ, Rpc: rpcReq}
+			err = ws.WriteJSON(req)
+			require.NoError(t, err)
 
-		var rsp map[string]interface{}
-		err = ws.ReadJSON(&rsp)
-		require.NoError(t, err)
-		require.Equal(t, eventbus.EVT_RPC_RSP, rsp["type"])
-		require.NotNil(t, rsp["rpc"])
+			var rsp map[string]interface{}
+			err = ws.ReadJSON(&rsp)
+			require.NoError(t, err)
+			require.Equal(t, eventbus.EVT_RPC_RSP, rsp["type"])
+			require.NotNil(t, rsp["rpc"])
 
-		rpcRsp := rsp["rpc"].(map[string]interface{})
-		require.Equal(t, "2.0", rpcRsp["jsonrpc"])
-		require.Equal(t, float64(id+1), rpcRsp["id"])
+			rpcRsp := rsp["rpc"].(map[string]interface{})
+			require.Equal(t, "2.0", rpcRsp["jsonrpc"])
+			require.Equal(t, float64(id+1), rpcRsp["id"])
 
-		if errObj, ok := rpcRsp["error"]; ok {
-			t.Logf("RPC Error: %v", errObj)
-			require.Fail(t, "RPC returned error")
-		} else {
-			result, ok := rpcRsp["result"]
-			require.True(t, ok, "RPC result not found")
-			require.Equal(t, tc.expect, result)
-		}
+			if errObj, ok := rpcRsp["error"]; ok {
+				t.Logf("RPC Error: %v", errObj)
+				require.Fail(t, "RPC returned error")
+			} else {
+				result, ok := rpcRsp["result"]
+				require.True(t, ok, "RPC result not found")
+				require.Equal(t, tc.expect, result)
+			}
+		})
 	}
 }
 
@@ -103,7 +121,7 @@ func TestWsLLMMessages(t *testing.T) {
 	}
 
 	// Ensure we have some LLM providers loaded
-	useTestingLLMProviders()
+	chat.SetTesting(true)
 
 	at, _, err := jwtLogin("sys", "manager")
 	require.Nil(t, err)
